% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption_functions.R
\name{abs_blcor}
\alias{abs_blcor}
\title{Baseline correction for absorbance data}
\usage{
abs_blcor(abs_data, wlrange = c(680, 700))
}
\arguments{
\item{abs_data}{data.frame containing samples in columns, the column containing wavelengths must be named "wavelength"}

\item{wlrange}{range of wavelengths that should be used for correction, absorbance mean in that range is subtracted from each value (sample-wise)}
}
\value{
data.frame
}
\description{
Baseline correction for absorbance data
}
\examples{
data(absorbance)
abs_data_cor <- abs_blcor(absorbance)

abs_data_cor1 <- abs_blcor(absorbance[1:2])

}
