% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{abs_data}
\alias{abs_data}
\title{raw data from absorbance measurements, 5cm pathlength}
\format{list of parafacs}
\usage{
abs_data
}
\description{
raw data from absorbance measurements, 5cm pathlength
}
\keyword{datasets}
