% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_importers.R
\name{.eem_csv}
\alias{.eem_csv}
\title{Import EEMs from generic csv files.}
\usage{
.eem_csv(file, col = "ex")
}
\arguments{
\item{file}{path to file}

\item{col}{either "ex" or "em", whatever wavelength is arranged in columns}
}
\value{
list with EEM data
}
\description{
Import EEMs from generic csv files.
}
