% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{.trans_parafac}
\alias{.trans_parafac}
\title{Add data of a PARAFAC model derived from multiway from EEMs}
\usage{
.trans_parafac(parafac, em, ex, samples, comp, const, norm_factors)
}
\arguments{
\item{parafac}{parafac model}

\item{em}{emission wavelengths}

\item{ex}{excitation wavelengths}

\item{samples}{sample names}

\item{comp}{number of components}

\item{const}{constraints}

\item{norm_factors}{factors to invert normalisation}
}
\value{
parafac model
}
\description{
Add data of a PARAFAC model derived from multiway from EEMs
}
