% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eem2array}
\alias{eem2array}
\title{Data from an eemlist is transformed into an array}
\usage{
eem2array(eem_list, normalise = FALSE)
}
\arguments{
\item{eem_list}{object of class eemlist}

\item{normalise}{logical value whether data should be normalised for each sample}
}
\value{
object of class array
}
\description{
Data matrices from EEM are combined to an array that is needed for a PARAFAC analysis.
}
\examples{
data(eem_list)

eem2array(eem_list)
}
