% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_checkdata}
\alias{eem_checkdata}
\title{Check your EEM, absorption and metadata before processing}
\usage{
eem_checkdata(eem_list, absorbance, metadata = NULL, metacolumns = NULL,
  error = TRUE)
}
\arguments{
\item{eem_list}{eemlist continaing EEM data.}

\item{absorbance}{data.frame containing absorbance data.}

\item{metadata}{optional data.frame containing metadata.}

\item{metacolumns}{character vector of columns that are checkt for complete data sets}

\item{error}{logical, whether a problem should cause an error or not.}
}
\value{
logical, whether severe errors occured
}
\description{
The function tries to lead you to possible problems in your data.
}
\examples{
data(eem_list)
data(abs_data)

eem_checkdata(eem_list,abs_data)
}
