% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_checksize}
\alias{eem_checksize}
\title{Check size of EEMs}
\usage{
eem_checksize(eem_list)
}
\arguments{
\item{eem_list}{eemlist}
}
\value{
character vector
}
\description{
The size of EEMs in an eemlist is checked and the sample names of samples with more data than the sample with the smallest range are returned.
}
\examples{
data(eem_list)
eem_checksize(eem_list)
}
