% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_corrections}
\alias{eem_corrections}
\title{Return names of samples where certain corrections are missing.}
\usage{
eem_corrections(eem_list)
}
\arguments{
\item{eem_list}{eemlist to be checked}
}
\value{
prints out sample names
}
\description{
Return names of samples where certain corrections are missing.
}
\examples{
data(eem_list)

eem_corrections(eem_list)
}
