% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_importers.R
\name{eem_csv2}
\alias{eem_csv2}
\title{Importer function for generic csv files to be used with eem_read().}
\usage{
eem_csv2(file)
}
\arguments{
\item{file}{path to file passed from eem_read}
}
\value{
list with EEM data
}
\description{
This function can be used to import generic csv files containing EEM data using \code{\link[eemR]{eem_read}}. Excitation wavelengths are assumed row-wise and emission wavelengths column-wise If your data is arranged the other way round, please use \code{\link{eem_csv}}
}
\examples{
\donttest{
## no example data provided with the package
## below is an example how this could like like
eems <- "C:/some/path/to/eem.csv"
eem_list <- eem_read(eems, recursive = TRUE, import_function = eem_csv2)

eem_list
}
}
