% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_duplicates}
\alias{eem_duplicates}
\alias{eem_duplicates.default}
\alias{eem_duplicates.eemlist}
\alias{eem_duplicates.data.frame}
\title{Check for duplicate sample names}
\usage{
eem_duplicates(data)

\method{eem_duplicates}{default}(data)

\method{eem_duplicates}{eemlist}(data)

\method{eem_duplicates}{data.frame}(data)
}
\arguments{
\item{data}{eemlist or data.frame containing absorbance data}
}
\value{
named character vector with duplicate sample names
}
\description{
Check for duplicate sample names
}
\examples{
### check
}
