% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_easy}
\alias{eem_easy}
\title{Opens an R markdown template for an easy and userfriendly analysis of EEM data.}
\usage{
eem_easy()
}
\value{
A pdf report, a peak picking table and optional plots.
}
\description{
In your default editor (e.g. RStudio), a Rmd file is opened. It consists of bloacks gathering the parameters and information needed and continues with a series of data corrections, peak picking and plots. Finally you get a report of your analysis, a table with the peaks and optional pngs of your fluorescence data. To continue working and keeping your settings, the file can be sa ved anywhere and reused anytime.
}
\details{
Function does not work well in Windows. You might try file.edit(system.file("EEM_simple_analysis.Rmd", package = "staRdom"))
}
\examples{
\dontrun{
#
eem_easy()

# this function fails very often, so you might use that:
file.edit(system.file("EEM_simple_analysis.Rmd", package = "staRdom"))
}
}
