% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_eemdil}
\alias{eem_eemdil}
\title{Correct names of EEM samples to match undiluted absorbance data.}
\usage{
eem_eemdil(eem_list, abs_data = NULL, dilution = NULL, cor_data = NULL,
  auto = TRUE, verbose = FALSE)
}
\arguments{
\item{eem_list}{eemlist}

\item{abs_data}{optinal absorbance data as data frame}

\item{dilution}{optinal dilution data as data frame}

\item{cor_data}{optional output from \code{\link{eem_dilcorr}} as data frame}

\item{auto}{optional, see \code{\link{eem_dilcorr}}}

\item{verbose}{optional, see \code{\link{eem_dilcorr}}}
}
\value{
eemlist
}
\description{
Correct names of EEM samples to match undiluted absorbance data.
}
\examples{
data(eem_list)
data(abs_data)

dilution <- data.frame(dilution = c(rep(1,10),rep(5,10),rep(10,10)))
rownames(dilution) <- eem_names(eem_list)

eem_list2 <- eem_eemdil(eem_list,abs_data,dilution)

}
