% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_importers.R
\name{eem_hitachi}
\alias{eem_hitachi}
\title{Importer function for Hitachi F-7000 txt files to be used with eem_read().}
\usage{
eem_hitachi(file)
}
\arguments{
\item{file}{path to file passed from eem_read}
}
\value{
list with EEM data
}
\description{
This function can be used to import txt files from Hitachi F-7000 containing EEM data using \code{\link[eemR]{eem_read}}.
}
\examples{
\donttest{
## no example data provided with the package
## below is an example how this could like like
# eems <- "C:/some/path/to/hitachi.TXT"
# eem_list <- eem_read(eems, recursive = TRUE, import_function = eem_hitachi)

# eem_list
}
}
