% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_import_dir}
\alias{eem_import_dir}
\title{Load all eemlist obects saved in different Rdata or RDa files in a folder.}
\usage{
eem_import_dir(dir)
}
\arguments{
\item{dir}{folder where RData files are saved}
}
\value{
eemlist
}
\description{
Reads Rdata and RDa files with one eemlist each. The eemlists are combined into one and returned.
}
\examples{
\dontrun{
# due to package size issues no example data is provided for this function
# eem_import_dir("C:/some_folder/with_EEMS/only_Rdata_files")
}
}
