% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_is.na}
\alias{eem_is.na}
\title{Check for NAs in EEM data}
\usage{
eem_is.na(eem_list)
}
\arguments{
\item{eem_list}{eemlist to check}
}
\value{
named character vector with sample names where EEM data contains NAs
}
\description{
Check for NAs in EEM data
}
\examples{
### check
}
