% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eem_list}
\alias{eem_list}
\title{15 fluorescence samples from drEEM used for examples.}
\format{
eemlist
}
\usage{
eem_list
}
\description{
15 fluorescence samples from drEEM used for examples.
}
\keyword{datasets}
