% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eem_list_667sf}
\alias{eem_list_667sf}
\title{Fluorescence data of 1 sample of the original data set. Used for comparison in the vignette.}
\format{eemlist}
\usage{
eem_list_667sf
}
\description{
Fluorescence data of 1 sample of the original data set. Used for comparison in the vignette.
}
\keyword{datasets}
