% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eem_list_outliers}
\alias{eem_list_outliers}
\title{2 fluorescence samples from drEEM that were excluded as outliers from the PARAFAC model.}
\format{
eemlist
}
\usage{
eem_list_outliers
}
\description{
2 fluorescence samples from drEEM that were excluded as outliers from the PARAFAC model.
}
\keyword{datasets}
