% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_load_dreem}
\alias{eem_load_dreem}
\title{Load original data from the drEEM tutorial and return it as eemlist}
\usage{
eem_load_dreem()
}
\value{
eemlist
}
\description{
Load original data from the drEEM tutorial and return it as eemlist
}
\examples{
\donttest{
# Reading MATLAB files from recent versions like the demo dataset from drEEM
# can cause problems if the R installation lacks UTF8 support in iconv.
# Therefore, we use try() in the example. If you encounter related problems,
# please refer to the help for R.matlab::readMat() for details.

eem_list <- try(eem_load_dreem(), silent = FALSE)
eem_list
}
}
