% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{eem_metatemplate}
\alias{eem_metatemplate}
\title{Create table that contains sample names and locations of files.}
\usage{
eem_metatemplate(eem_list = NULL, absorbance = NULL)
}
\arguments{
\item{eem_list}{eemlist}

\item{absorbance}{data frame with absorbance data}
}
\value{
data frame
}
\description{
You can use this table as an overview of your files and/or as a template for creating a metadata table.
}
\examples{
folder <- system.file("extdata/EEMs", package = "staRdom") # load example data
eem_list <- eem_read_csv(folder)
data(absorbance)

eem_metatemplate(eem_list,absorbance)
}
