% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eem_parafac}
\alias{eem_parafac}
\title{Runs a PARAFAC analysis on EEM data}
\usage{
eem_parafac(eem_list, comps, maxit = 500, normalise = TRUE,
  const = c("nonneg", "nonneg", "nonneg"), nstart = 10, ctol = 10^-4,
  cores = parallel::detectCores()/2, verbose = FALSE, ...)
}
\arguments{
\item{eem_list}{object of class \code{\link[eemR]{eem}}}

\item{comps}{vector containing the desired numbers of components. For each of these numbers one model is calculated}

\item{maxit}{maximum iterations for PARAFAC algorithm}

\item{normalise}{state whether EEM data should be normalised in advance}

\item{const}{constraints of PARAFAC analysis. Default is non-negative ("nonneg"), alternatively smooth and non-negative ("smonon") might be interesting for an EEM analysis.}

\item{nstart}{number of random starts}

\item{ctol}{Convergence tolerance (R^2 change)}

\item{cores}{number of parallel calculations (e.g. number of physical cores in CPU)}

\item{verbose}{print infos}

\item{...}{additional parameters that are passed on to \code{\link[multiway]{parafac}}}
}
\value{
object of class parafac
}
\description{
One or more PARAFAC models can be calculated depending on the number of components. The idea is to compare the different models to get the most suitable. B-mode is emmission wavelengths, C-mode is excitation wavelengths and, A-mode is the loadings of the samples. The calculation is done with \code{\link[multiway]{parafac}}, please see details there.
}
\examples{
\donttest{
data(eem_list)

dim_min <- 3 # minimum number of components
dim_max <- 7 # maximum number of components
nstart <- 10 # random starts for PARAFAC analysis, models built simulanuously, best selected
cores <- parallel::detectCores()/2 # use all cores but do not use all threads
maxit = 500
ctol <- 10^-4 # tolerance for parafac

pfres_comps <- eem_parafac(eem_list,comps=seq(dim_min,dim_max),
    normalise = TRUE,maxit=10000,nstart=nstart,ctol=ctol,cores=cores)
}
}
\seealso{
\code{\link[multiway]{parafac}}
}
