% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_scale_ext}
\alias{eem_scale_ext}
\title{Determine the range of fluorescence values in a set of samples}
\usage{
eem_scale_ext(data)
}
\arguments{
\item{data}{eemlist containing the EEM data}
}
\value{
numeric vector
}
\description{
Determine the range of fluorescence values in a set of samples
}
\examples{
data(eem_list)
eem_scale_ext(eem_list)
}
