% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction_functions.R
\name{eem_smooth}
\alias{eem_smooth}
\title{Smooth fluorescence data by calculating rolling mean along excitation wavelengths.}
\usage{
eem_smooth(data, n = 4)
}
\arguments{
\item{data}{fluorescence data of class eemlist}

\item{n}{width of rolling mean window in nm}
}
\value{
eemlist with smoothed data
}
\description{
Smooth fluorescence data by calculating rolling mean along excitation wavelengths.
}
\examples{
data(eem_list)

eem_list <- eem_smooth(eem_list,n=4)
}
