% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_write_csv}
\alias{eem_write_csv}
\title{Export samples in an EEM list to a single csv files}
\usage{
eem_write_csv(eem_list, output, ...)
}
\arguments{
\item{eem_list}{EEM data as eemlist}

\item{output}{path to folder where csv files are exported to}

\item{...}{additional arguments}
}
\value{
returns the exported EEMs as a list of data.frames
}
\description{
Export samples in an EEM list to a single csv files
}
\examples{
data(eem_list)

output <- tempdir()
output
a <- eem_write_csv(eem_list, output)
}
