% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_bindxc}
\alias{eempf_bindxc}
\title{Combining extracted components of PARAFAC models}
\usage{
eempf_bindxc(components)
}
\arguments{
\item{components}{list of parafac_components}
}
\value{
parafac_components
}
\description{
Combining extracted components of PARAFAC models
}
\examples{

data(pf_models)
pfmodel <- pf4[[1]]
comps <- eempf_excomp(pfmodel,c(1,3))
comps2 <- eempf_excomp(pfmodel,c(4,6))
comps3 <- eempf_bindxc(list(comps, comps2))

}
