% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_C_plot}
\alias{eempf_C_plot}
\title{Plot amount of each component in each sample as bar plot}
\usage{
eempf_C_plot(cp_out)
}
\arguments{
\item{cp_out}{parafac model}
}
\value{
ggplot
}
\description{
Plot amount of each component in each sample as bar plot
}
\examples{
\dontrun{
data(pfres_comps1)

eempf_C_plot(pfres_comps[[2]][[3]])
}
}
