% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_comp_C_plot}
\alias{eempf_comp_C_plot}
\title{Plot components from a PARAFAC model}
\usage{
eempf_comp_C_plot(cp_out)
}
\arguments{
\item{cp_out}{object of class parafac}
}
\value{
ggplot
}
\description{
Additionally a bar plot with the amounts of each component in each sample is produced.
}
\examples{
\dontrun{
data(pfres_comps1)

eempf_comp_C_plot(pfres_comps[[2]][[3]])
}
}
\seealso{
\code{\link[staRdom]{ggeem}}, \code{\link[staRdom]{eempf_C_plot}}
}
