% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_comp_load_plot}
\alias{eempf_comp_load_plot}
\title{Plot components from a PARAFAC model}
\usage{
eempf_comp_load_plot(pfmodel)
}
\arguments{
\item{pfmodel}{object of class parafac}
}
\value{
ggplot
}
\description{
Additionally a bar plot with the amounts of each component in each sample is produced.
}
\examples{
data(pf_models)

eempf_comp_load_plot(pf4[[1]])
}
\seealso{
\code{\link[staRdom]{ggeem}}, \code{\link[staRdom]{eempf_load_plot}}
}
