% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_comp_mat}
\alias{eempf_comp_mat}
\title{Extract EEM matrix for single components determined in the PARAFAC analysis}
\usage{
eempf_comp_mat(pfmodel, gather = TRUE)
}
\arguments{
\item{pfmodel}{object of class parafac}

\item{gather}{logical value whether excitation wavelengths are a column, otherwise excitation wavelengths are column names}
}
\value{
a list of class data frames
}
\description{
The components of a PARAFAC analysis are extracted as a data frame
}
\examples{
data(pf_models)

eempf_comp_mat(pf4[[1]])
}
