% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_comp_names<-}
\alias{eempf_comp_names<-}
\title{Set names of PARAFAC components}
\usage{
eempf_comp_names(pfmodel) <- value
}
\arguments{
\item{pfmodel}{model of class parafac}

\item{value}{character vector containing the new names for the components}
}
\value{
parafac model
}
\description{
Set names of PARAFAC components
}
\examples{
data(pf_models)

eempf_comp_names(pf4) <- c("A","B","C","D","E","F","G")
}
