% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_comp_names}
\alias{eempf_comp_names}
\title{Extract names from PARAFAC model components}
\usage{
eempf_comp_names(pfmodel)
}
\arguments{
\item{pfmodel}{parafac model}
}
\value{
vector of names or list of vecters of names
}
\description{
Extract names from PARAFAC model components
}
\examples{
data(pf_models)
eempf_comp_names(pf4)

eempf_comp_names(pf4) <- c("A","B","C","D","E","F","G")

value <- list(c("A1","B1","C1","D","E","F","G"),
c("A2","B2","C","D","E","F","G"),
c("A3","B3","C","D","E","F","G"),
c("A4","B4","C","D","E","F","G"),
c("A5","B5","C","D","E","F","G5")
)

eempf_comp_names(pf4) <- value
eempf_comp_names(pf4)

ggeem(pf4[[1]])

}
