% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_compare}
\alias{eempf_compare}
\title{Plot a set of PARAFAC models to compare the single components}
\usage{
eempf_compare(pfres, ...)
}
\arguments{
\item{pfres}{list of several objects of class parafac}

\item{...}{arguments passed on to \code{\link[staRdom]{eempf_fits}} and \code{\link[staRdom]{eempf_plot_comps}}}
}
\value{
3 objects of class ggplot
}
\description{
Three plots are returned:
\enumerate{
    \item plot of numer of components vs. model fit
    \item plot of different components as colour maps
    \item plot of different components as peak lines
}
The plots are intended to help with a suitable number of components.
}
\examples{
\donttest{
data(pf_models)

eempf_compare(pf4)
}
}
\seealso{
\code{\link[staRdom]{eempf_fits}}, \code{\link[staRdom]{eempf_plot_comps}}
}
