% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_comps3D}
\alias{eempf_comps3D}
\title{3D plots of PARAFAC components}
\usage{
eempf_comps3D(pfmodel, which = NULL)
}
\arguments{
\item{pfmodel}{object of class parafac}

\item{which}{optional, if numeric selects certain component}
}
\value{
plotly plot
}
\description{
Interactive 3D plots are created using plotly.
}
\examples{
\dontrun{
data(pf_models)

eempf_comps3D(pf4[[1]])
}
}
