% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_convergence}
\alias{eempf_convergence}
\title{Extract modelling imformation from a PARAFAC model.}
\usage{
eempf_convergence(pfmodel, print = TRUE)
}
\arguments{
\item{pfmodel}{PARAFAC model created with staRdom using output = "all"}

\item{print}{logical, whether you want console output or just a list with results}
}
\value{
list with numbers of converging models, cflags and SSEs
}
\description{
The convergence behaviour of all initialisations in a PARAFAC model is shown by printing the numbers
}
\examples{
data("pf_models")

pfmodel <- pf4[[1]]
conv_beh <- eempf_convergence(pfmodel)
}
