% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_corcondia}
\alias{eempf_corcondia}
\title{Calculate the core consistancy of an EEM PARAFAC model}
\usage{
eempf_corcondia(pfmodel, eem_list, divisor = "core")
}
\arguments{
\item{pfmodel}{PARAFAC model}

\item{eem_list}{eemlist}

\item{divisor}{divisor, please refer to \code{\link[multiway]{corcondia}}}
}
\value{
numeric
}
\description{
This is basically a wrapper for \code{\link[multiway]{corcondia}} that deals with the normalisation of the original data., Other than \code{\link[multiway]{corcondia}}, the default dicisor = "core".
}
\examples{
\dontrun{
# due to data limitation in package, example does not work with that data!

# eempf_corcondia(pfmodel,eem_list)

}
}
