% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_corplot}
\alias{eempf_corplot}
\title{Plot correlations of components in samples}
\usage{
eempf_corplot(pfmodel, normalisation = FALSE, lower = list(continuous =
  "smooth"), mapping = aes(alpha = 0.2), ...)
}
\arguments{
\item{pfmodel}{object of class parafac}

\item{normalisation}{logical, whether normalisation is undone or not}

\item{lower}{style of lower plots, see \code{\link[GGally]{ggpairs}}}

\item{mapping}{aesthetic mapping, see \code{\link[GGally]{ggpairs}}}

\item{...}{passed on to \code{\link[GGally]{ggpairs}}}
}
\value{
object of class ggplot
}
\description{
A pair plot showing correlations between samples is created.
}
\examples{
\donttest{
data(pf_models)
eempf_corplot(pf4[[1]])
}

}
\seealso{
\code{\link[GGally]{ggpairs}}
}
