% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_cortable}
\alias{eempf_cortable}
\title{Calculating correlations between the component loadings in all samples (C-Modes).}
\usage{
eempf_cortable(cp_out, method = "pearson", ...)
}
\arguments{
\item{cp_out}{results from a PARAFAC analysis, class parafac}

\item{method}{method of correlation, passed to \code{\link[stats]{cor}}}

\item{...}{passed on to \code{\link[stats]{cor}}}
}
\value{
matrix
}
\description{
Calculating correlations between the component loadings in all samples (C-Modes).
}
\examples{
data(pfres_comps1)
eempf_cortable(pfres_comps[[2]])
}
