% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_eemqual}
\alias{eempf_eemqual}
\title{Calculating EEMqual which is an indicator of a PARAFAC model's quality}
\usage{
eempf_eemqual(eem_list, cp_out, ...)
}
\arguments{
\item{eem_list}{EEM data as eemlist}

\item{cp_out}{PARAFAC model}

\item{...}{additional arguments passed to splithalf()}
}
\value{
list containing eemqual and splithalf results
}
\description{
Calculating EEMqual which is an indicator of a PARAFAC model's quality
}
\examples{
\dontrun{
data(eem_list)
data(pfres_comps1)

eempf_eemqual(eem_list,pfres_comps[[3]])
}
}
\references{
Rasmus Bro, Maider Vidal, EEMizer: Automated modeling of fluorescence EEM data, Chemometrics and Intelligent Laboratory Systems, Volume 106, Issue 1, 2011, Pages 86-92, ISSN 0169-7439
}
