% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_export}
\alias{eempf_export}
\title{Create one table containing the PARAFAC models factors and optionally exporting it to csv or txt}
\usage{
eempf_export(pfmodel, export = NULL, Fmax = TRUE, ...)
}
\arguments{
\item{pfmodel}{PARAFAC model}

\item{export}{file path to export table}

\item{Fmax}{rescale modes so the A mode shows the maximum fluorescence}

\item{...}{additional parameters passed to \code{\link[utils]{write.table}}}
}
\value{
data frame
}
\description{
Create one table containing the PARAFAC models factors and optionally exporting it to csv or txt
}
\examples{
data(pf_models)

factor_table <- eempf_export(pf4[[1]])
}
