% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_fits}
\alias{eempf_fits}
\title{Fits vs. components of PARAFAC models are plotted}
\usage{
eempf_fits(pfres)
}
\arguments{
\item{pfres}{list of objects of class parafac}
}
\value{
object of class ggplot
}
\description{
Fits vs. components of PARAFAC models are plotted
}
\examples{
data(pf_models)

eempf_fits(pf4)
}
