% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_leverage}
\alias{eempf_leverage}
\title{Calculate the leverage of each emission and excitation wavelength and each sample}
\usage{
eempf_leverage(cp_out)
}
\arguments{
\item{cp_out}{object of class parafac}
}
\value{
list of 3 named vectors (emission, excitation wavelengths and samples)
}
\description{
Calculate the leverage of each emission and excitation wavelength and each sample
}
\examples{
data(pfres_comps1)

eempf_leverage(pfres_comps[[2]])
}
