% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_leverage_ident}
\alias{eempf_leverage_ident}
\title{Plot leverage of emission wavelengths, excitation wavelengths and samples.}
\usage{
eempf_leverage_ident(cpl, qlabel = 0.1)
}
\arguments{
\item{cpl}{leverage, outpout from \code{\link[staRdom]{eempf_leverage}}}

\item{qlabel}{optional, quantile of which labels are shown (1 = all, 0 = no labels)}
}
\value{
list of three vectors containing the names of selected samples
}
\description{
Plot is interactive where you can select values with your mouse. A list of vectors is returned to remove this outliers in a further step from your samples. The labels to be shown can be selected by adding the quatile of samples with highest leverages to be labeled.
}
\examples{
data(pfres_comps1)

leverage <- eempf_leverage(pfres_comps[[2]])
outliers <- eempf_leverage_ident(leverage)
}
\seealso{
\code{\link[staRdom]{eempf_leverage_plot}}
}
