% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_leverage_plot}
\alias{eempf_leverage_plot}
\title{Plot leverage of emission wavelengths, excitation wavelengths and samples.}
\usage{
eempf_leverage_plot(cpl, qlabel = 0.1)
}
\arguments{
\item{cpl}{leverage, outpout from \code{\link[staRdom]{eempf_leverage}}}

\item{qlabel}{optional, quantile of which labels are shown (1 = all, 0 = no labels)}
}
\value{
ggplot
}
\description{
The labels to be shown can be selected by adding the quatile of samples with highest leverages to be labeled.
}
\examples{
data(pf_models)

leverage <- eempf_leverage(pf4[[1]])
eempf_leverage_plot(leverage)
}
\seealso{
\code{\link[staRdom]{eempf_leverage_ident}}
}
