% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_load_plot}
\alias{eempf_load_plot}
\title{Plot amount of each component in each sample as bar plot}
\usage{
eempf_load_plot(pfmodel)
}
\arguments{
\item{pfmodel}{parafac model}
}
\value{
ggplot
}
\description{
Plot amount of each component in each sample as bar plot
}
\examples{
data(pfres_comps1)

eempf_load_plot(pfres_comps[[2]])
}
