% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_mleverage}
\alias{eempf_mleverage}
\title{Calculate the leverage of each emission and excitation wavelength and each sample from a list of PARAFAC models}
\usage{
eempf_mleverage(pfres_comps, ecdf = FALSE, stats = FALSE)
}
\arguments{
\item{pfres_comps}{object of class parafac}

\item{ecdf}{logical, transforme leverages to according empirical quantiles (\code{\link[stats]{ecdf}})}

\item{stats}{logical, whether means and standard deviations are calculated from leverages}
}
\value{
data frame containing leverages of wavelengths and samples for each model
}
\description{
Calculate the leverage of each emission and excitation wavelength and each sample from a list of PARAFAC models
}
\examples{
data(pfres_comps1)

eempf_mleverage(pfres_comps)
}
