% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_plot_comps}
\alias{eempf_plot_comps}
\title{Plot all components of PARAFAC models}
\usage{
eempf_plot_comps(pfres, type = 1)
}
\arguments{
\item{pfres}{list of PARAFAC models}

\item{type}{1 for a colour map and 2 for peak lines}
}
\value{
object of class ggplot
}
\description{
The components can be plottet in two ways: either as a colour map or as two lines (emission, excitation wavelengths) intersecting at the component maximum.
}
\examples{
data(pfres_comps1)

eempf_plot_comps(pfres_comps)
eempf_plot_comps(pfres_comps,type=2)
}
