% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_reorder}
\alias{eempf_reorder}
\title{Reorder PARAFAC components}
\usage{
eempf_reorder(pfmodel, order, decreasing = FALSE)
}
\arguments{
\item{pfmodel}{model of class parafac}

\item{order}{vector containing desired new order or "em" or "ex" to reorder according to emission or excitation wavelengths of the peaks}

\item{decreasing}{logical, whether components are reordered according to peak wvalengths in a decreasing direction}
}
\value{
parafac model
}
\description{
Reorder PARAFAC components
}
\examples{
data(pf_models)
ggeem(pf4[[1]])

pf4r <- eempf_reorder(pf4[[1]],"ex")
ggeem(pf4r)
}
