% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_rescaleAB}
\alias{eempf_rescaleAB}
\title{Rescale A and B modes of PARAFAC model}
\usage{
eempf_rescaleAB(pfmodel, newscale = 1)
}
\arguments{
\item{pfmodel}{object of class parafac}

\item{newscale}{Desired root mean-square for each column of rescaled mode. Can input a scalar or a vector with length equal to the number of factors for the given mode. If newscale = "Fmax", each component will be scaled so the maximum of each component is 1.}
}
\value{
object of class parafac
}
\description{
B and C modes (emission and excitation wavelengths) are rescaled to RMS of value newscale. This is compensated in A mode (sample loadings).
}
\examples{
data(pfres_comps1)

new_pf <- eempf_rescaleAB(pfres_comps[[2]])
}
\seealso{
\code{\link[multiway]{rescale}}
}
