% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_residuals}
\alias{eempf_residuals}
\title{Calculate residuals of EEM data according to a certain model}
\usage{
eempf_residuals(
  pfmodel,
  eem_list,
  select = NULL,
  cores = parallel::detectCores(logical = FALSE)/2
)
}
\arguments{
\item{pfmodel}{PARAFAC model of class parafac}

\item{eem_list}{eemlist containing EEM data}

\item{select}{character vector containing the names of the desired samples}

\item{cores}{number of cores to use for parallel processing}
}
\value{
data frame with EEM residuals
}
\description{
Calculate residuals of EEM data according to a certain model
}
\examples{
\donttest{
data(eem_list)
data(pf_models)

eempf_residuals(pf4[[1]],eem_list)
}
}
