% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{maxlines}
\alias{maxlines}
\title{Extract data from emission and excitation wavelengths of the components of a PARAFAC model}
\usage{
maxlines(cp_out)
}
\arguments{
\item{cp_out}{object of class parafac}
}
\value{
data frame
}
\description{
Data of wavelengths is returned. For each component the lines intersecting at the component maxima are returned.
}
\examples{
data(pfres_comps1)

ml <- maxlines(pfres_comps[[1]])
}
