% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{norm2C}
\alias{norm2C}
\title{Compensate for normalisation in C-modes}
\usage{
norm2C(cp_out)
}
\arguments{
\item{cp_out}{object of class parafac}
}
\value{
object of class parafac
}
\description{
Factors used for normalisation are saved separately in the PARAFAC models. With this function, the normalisation factors are combined with the C-modes of the model and removed as a separate vector. This means former normalisation is accounted for in the amount of each component in each sample. If no normalisation was done, the original model is returned without warning.
}
\examples{
\dontrun{
data(pfres_comps1)

pfres_comps[[2]][[3]] <- norm2C(pfres_comps[[2]][[3]])
}
}
