% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{norm_array}
\alias{norm_array}
\title{Normalise 3-dimensional array in first and second dimension}
\usage{
norm_array(eem_array)
}
\arguments{
\item{eem_array}{3-dimensional array}
}
\value{
array
}
\description{
Normalise 3-dimensional array in first and second dimension
}
\examples{
data(eem_list)

a <- eem2array(eem_list)
an <- norm_array(a)
}
