% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{peaks}
\alias{peaks}
\title{data.frame containing common EEM peaks with the wavelengths and possible deviations}
\format{
data.frame
}
\usage{
peaks
}
\description{
data.frame containing common EEM peaks with the wavelengths and possible deviations
}
\keyword{datasets}
