% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pf1n}
\alias{pf1n}
\title{PARAFAC model, see vignette, non-negative constraints}
\format{
list of parafacs
}
\usage{
pf1n
}
\description{
PARAFAC model, see vignette, non-negative constraints
}
\keyword{datasets}
