% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pf2}
\alias{pf2}
\title{PARAFAC model, see vignette, non-negative constraints, normalised}
\format{
list of parafacs
}
\usage{
pf2
}
\description{
PARAFAC model, see vignette, non-negative constraints, normalised
}
\keyword{datasets}
