% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pf3}
\alias{pf3}
\title{PARAFAC model, see vignette, non-negative constraints, normalised, outliers removed}
\format{
list of parafacs
}
\usage{
pf3
}
\description{
PARAFAC model, see vignette, non-negative constraints, normalised, outliers removed
}
\keyword{datasets}
