% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pf4}
\alias{pf4}
\title{PARAFAC model, see vignette, non-negative constraints, normalised, outliers removed, high accuarcy}
\format{list of parafacs}
\usage{
pf4
}
\description{
PARAFAC model, see vignette, non-negative constraints, normalised, outliers removed, high accuarcy
}
\keyword{datasets}
