% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sh}
\alias{sh}
\title{result from PARAFAC split-half analysis, periodic data split}
\format{
list of parafacs
}
\usage{
sh
}
\description{
result from PARAFAC split-half analysis, periodic data split
}
\keyword{datasets}
