% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sh_r}
\alias{sh_r}
\title{result from PARAFAC split-half analysis, random data split}
\format{list of parafacs}
\usage{
sh_r
}
\description{
result from PARAFAC split-half analysis, random data split
}
\keyword{datasets}
