% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{splithalf_plot}
\alias{splithalf_plot}
\title{Plot results from a splithalf analysis}
\usage{
splithalf_plot(fits)
}
\arguments{
\item{fits}{list of components data}
}
\value{
ggplot
}
\description{
Graphs of all components of all models are plotted to be compared.
}
\examples{
data(sh)

splithalf_plot(sh)
}
\seealso{
\code{\link[staRdom]{splithalf}}
}
