% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{ssc_max}
\alias{ssc_max}
\title{Calculate the combination of components giving the maximum of geometric mean of TCCs}
\usage{
ssc_max(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
vector with TCCs having the highest possible geometric mean
}
\description{
Calculate the combination of components giving the maximum of geometric mean of TCCs
}
\examples{
mat <- matrix(c(7,2,13,6,0,7,1,5,5), nrow = 3)
mat

sscs <- ssc_max(mat)
sscs

# order of components:
attr(sscs,"order")
}
