% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{tcc}
\alias{tcc}
\title{Caluclate Tucker's Congruence Coefficient of PARAFAC components}
\usage{
tcc(maxl_table, na.action = "na.omit")
}
\arguments{
\item{maxl_table}{data frame containing the peak lines of components}

\item{na.action}{if "na.omit" NA are deleted from prior the test}
}
\value{
data.frame containing the TCCs
}
\description{
Componets must be passed as modes, see \code{\link[staRdom]{maxlines}}
}
\examples{
data(pf_models)

ml <- maxlines(pf4[[1]])

tcc(ml)
}
