% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eem_list}
\alias{eem_list}
\title{Fluorescence data of 52 corrected samples.}
\format{eemlist}
\usage{
eem_list
}
\description{
Fluorescence data of 52 corrected samples.
}
\keyword{datasets}
