% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_load_dreem}
\alias{eem_load_dreem}
\title{Load original data from the drEEM tutorial and return it as eemlist}
\usage{
eem_load_dreem()
}
\value{
eemlist
}
\description{
Load original data from the drEEM tutorial and return it as eemlist
}
\examples{
\donttest{
eem_list <- eem_load_dreem()
}
}
