% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{eem_overview_plot}
\alias{eem_overview_plot}
\title{Plot fluorescence data from several samples split into several plots.}
\usage{
eem_overview_plot(data, spp = 8)
}
\arguments{
\item{data}{fluorescence data of class eemlist}

\item{spp}{number of samples per plot}
}
\value{
list of ggplots
}
\description{
Plot fluorescence data from several samples split into several plots.
}
\examples{
\donttest{
data(eem_list)
eem_overview_plot(eem_list,spp=3)
}
}
