% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{eem_red2smallest}
\alias{eem_red2smallest}
\title{Reduce wavelength range of EEM spectra to widest available in the whole sample set.}
\usage{
eem_red2smallest(data, verbose = FALSE)
}
\arguments{
\item{data}{data of EEM samples as eemlist}

\item{verbose}{states whether additional information is given in the command line}
}
\value{
eemlist with reduced spectral width
}
\description{
Reduce wavelength range of EEM spectra to widest available in the whole sample set.
}
\details{
This step is neccessary to perform a PARAFAC analysis which can only be calculated with spectra of similar range.
}
\examples{
data(eem_list)
eem_red2smallest(eem_list)
}
