% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_cortable}
\alias{eempf_cortable}
\title{Calculating correlations between the component loadings in all samples (C-Modes).}
\usage{
eempf_cortable(pfmodel, normalisation = FALSE, method = "pearson", ...)
}
\arguments{
\item{pfmodel}{results from a PARAFAC analysis, class parafac}

\item{normalisation}{logical, whether normalisation is undone or not}

\item{method}{method of correlation, passed to \code{\link[stats]{cor}}}

\item{...}{passed on to \code{\link[stats]{cor}}}
}
\value{
matrix
}
\description{
Calculating correlations between the component loadings in all samples (C-Modes).
}
\examples{
data(pf_models)
eempf_cortable(pf4[[1]])
}
