% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{eempf_leverage}
\alias{eempf_leverage}
\title{Calculate the leverage of each emission and excitation wavelength and each sample from a single PARAFAC model}
\usage{
eempf_leverage(pfmodel)
}
\arguments{
\item{pfmodel}{object of class parafac}
}
\value{
list of 3 named vectors (emission, excitation wavelengths and samples)
}
\description{
Calculate the leverage of each emission and excitation wavelength and each sample from a single PARAFAC model
}
\examples{
data(pfres_comps1)

eempf_leverage(pfres_comps[[2]])
}
