% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_plot_functions.R
\name{eempf_plot_ssccheck}
\alias{eempf_plot_ssccheck}
\title{Plot results from an SSC check}
\usage{
eempf_plot_ssccheck(ssccheck)
}
\arguments{
\item{ssccheck}{outpout from \code{\link{eempf_ssccheck}}}
}
\value{
ggplot element
}
\description{
Plot results from an SSC check
}
\examples{
data(pf_models)

ssccheck <- eempf_ssccheck(pf3[1:3])
eempf_plot_ssccheck(ssccheck)
}
