% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac_functions.R
\name{maxlines}
\alias{maxlines}
\title{Extract data from emission and excitation wavelengths of the components of a PARAFAC model (scaled B- and C-modes)}
\usage{
maxlines(pfmodel)
}
\arguments{
\item{pfmodel}{object of class parafac}
}
\value{
data frame
}
\description{
Data for each wavelengths is returned. For each component the lines intersecting at the component maxima are returned.
}
\examples{
data(pf_models)

ml <- maxlines(pf4[[1]])
}
